Move CubeGeometry to new meshes directory
